requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.2
django-jazzmin>=3.0.0
django-cors-headers>=4.4
gunicorn>=21.2
whitenoise>=6.7
